refactor: extract IssueExtractionRequestBudget.swift
Extracts the public IssueExtractionRequestBudget enum (67 lines) from
IssueExtractionRequestBuilder.swift into its own file.

Byte-identical move; SHA-256 afce7f11ad01a9354c6deb6e81367b6d82cd3e14035dcdbf211869df0d816b53
verified against origin/main:IssueExtractionRequestBuilder.swift lines 204-270.

Reduces IssueExtractionRequestBuilder.swift from 325 to 256 lines.

The remaining private encoding types (ChatCompletionRequest, ChatMessage,
ChatMessageContent, ChatMessageInputPart, ResponseFormat, ImageURLPayload)
stay in the builder file — they're all `private` (file-scope) helpers used
internally by IssueExtractionRequestBuilder to construct API payloads. Moving
them would require visibility widening.

Tests: 667 (unchanged).
